fix: prevent reuse of user IDs
Previously user IDs still used MAX(id) plus one while submission IDs used
SQLite allocation. This commit applies the same non-reuse invariant to
users and migrates both ID tables through one script.

- Allocate user IDs through SQLite and persist them in the JSON payload
- Create fresh user tables with AUTOINCREMENT
- Rename and extend the migration to convert users and submissions
