+++ server/db.py
@@ -68,16 +68,17 @@ async def delete_user(uid: int) -> None:
 async def create_user(user: dict) -> int:
     async with _open_db() as db:
         await db.execute("BEGIN EXCLUSIVE")
-        async with db.execute("SELECT MAX(id) FROM users") as cur:
-            row = await cur.fetchone()
-            if row is None:
-                raise RuntimeError("Failed to fetch max user ID.")
-            new_id = (row[0] or 0) + 1
-            
+        async with db.execute(
+            "INSERT INTO users (data) VALUES ('{}')"
+        ) as cur:
+            new_id = cur.lastrowid
+        if new_id is None:
+            raise RuntimeError("Failed to create user.")
+
         user["id"] = new_id
         await db.execute(
-            "INSERT INTO users (id, data) VALUES (?, ?)",
-            (new_id, json.dumps(user)),
+            "UPDATE users SET data = ? WHERE id = ?",
+            (json.dumps(user), new_id),
         )
         await db.commit()
         return new_id
@@ -247,7 +248,8 @@ async def init_db() -> None:
         await db.execute("PRAGMA journal_mode=WAL;")
         await db.execute("PRAGMA busy_timeout=15000;")
         await db.execute(
-            "CREATE TABLE IF NOT EXISTS users (id INTEGER PRIMARY KEY, data TEXT NOT NULL)"
+            "CREATE TABLE IF NOT EXISTS users "
+            "(id INTEGER PRIMARY KEY AUTOINCREMENT, data TEXT NOT NULL)"
         )
         await db.execute(
             "CREATE TABLE IF NOT EXISTS submissions "
